--- app/services/match-fetching.js
@@ -1,4 +1,0 @@
-import Service from '@ember/service';
-
-export default Service.extend({
-});
