+++ app/templates/components/live-matches.hbs
@@ -2,8 +2,7 @@
   {{#each matches as |match|}}
     {{live-match
       match=match
-      clockTime=clockTime
-      fetchMatches=fetchMatches
+      matchService=matchService
     }}
   {{/each}}
 </div>
