+++ app/templates/components/match-list.hbs
@@ -12,7 +12,7 @@
   {{/each}}
 </div>
 
-{{#if initialMatchFetch.isRunning}}
+{{#if matchService.initialFetchRunning}}
   <div class="layout__box o__flexes-to-1 o__centers-horizontally">
     <div class="t__centered">
       <div class="loader"></div>
