+++ app/templates/sweepstake.hbs
@@ -1,6 +1,6 @@
 <div class="layout__box o__flexes-to-1 o__has-rows main-content">
   {{pool-table players=model filterPlayer=filterPlayer toggleFilter=(action 'toggleFilter')}}
-  {{live-matches players=model fetchTask=fetchTask}}
-  {{past-matches players=model filterPlayer=filterPlayer fetchTask=fetchTask}}
-  {{match-list players=model filterPlayer=filterPlayer fetchTask=fetchTask}}
+  {{live-matches players=model}}
+  {{past-matches players=model filterPlayer=filterPlayer}}
+  {{match-list players=model filterPlayer=filterPlayer}}
 </div>
